Begin layout and customization separation
Splits the rendering/layout logic from the grid selection logic. The
goal is to use the same rendering engine for `<Group>` components.
